refactor(kernel-api): lift ProtocolParameterStore types and trait to kernel-api
Move protocol parameter types (ProtocolParameter, ParameterValue,
ParameterScope, ParameterConstraints, ParameterChange, PendingParameterChange)
and the ProtocolParameterStore trait from icn-governance to icn-kernel-api.

Governance now re-exports from kernel-api, so SledParameterStore and
InMemoryParameterStore directly implement the kernel-api trait. This lets
icn-core reference protocol parameter types without importing icn-governance.

Key changes:
- ParameterScope uses String instead of EntityId (kernel-api stays domain-free)
- get_effective() signature: Option<&EntityId> → Option<&str>
- icn-core refs switched: background_tasks, lifecycle, actor ParameterValue
- Governance ratchet: 83 → 75 (-8)
- kernel-api has zero icn_governance references
