+++ icn/crates/icn-governance/src/protocol.rs
@@ -48,639 +48,16 @@
 //! ).with_scope(ParameterScope::Global);
 //! ```
 
-use icn_entity::EntityId;
 use serde::{Deserialize, Serialize};
-use std::fmt;
 
-/// Known parameter categories for validation.
-///
-/// These are the standard categories used in ICN protocol parameters.
-/// New categories can be added but should be added here for validation.
-pub const KNOWN_PARAMETER_CATEGORIES: &[&str] = &[
-    "gossip",
-    "network",
-    "ledger",
-    "governance",
-    "trust",
-    "ratelimit",
-    "compute",
-];
-
-// ============================================================================
-// ProtocolParameter
-// ============================================================================
-
-/// A governable protocol parameter
-///
-/// Protocol parameters are configuration values that can be changed through
-/// the governance process. Each parameter has:
-/// - A unique identifier (e.g., "gossip.max_message_size")
-/// - Constraints on valid values
-/// - A scope determining where it applies
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProtocolParameter {
-    /// Unique parameter identifier (e.g., "gossip.max_message_size")
-    ///
-    /// Format: `<category>.<name>` or `<category>.<subcategory>.<name>`
-    pub id: String,
-
-    /// Human-readable name
-    pub name: String,
-
-    /// Detailed description of what this parameter controls
-    pub description: String,
-
-    /// Current value
-    pub value: ParameterValue,
-
-    /// Constraints on valid values
-    pub constraints: ParameterConstraints,
-
-    /// Scope where this parameter applies
-    pub scope: ParameterScope,
-
-    /// When the parameter was last updated (Unix timestamp)
-    pub updated_at: u64,
-
-    /// Proposal ID that last changed this parameter (if any)
-    pub updated_by: Option<String>,
-
-    /// Version for optimistic locking
-    ///
-    /// Incremented on each update to detect concurrent modifications.
-    /// When updating, the caller should provide the version they read;
-    /// if it doesn't match the stored version, the update is rejected.
-    #[serde(default)]
-    pub version: u64,
-}
-
-impl ProtocolParameter {
-    /// Validate that a parameter ID follows the required format.
-    ///
-    /// Valid formats:
-    /// - `category.name` (e.g., "gossip.fanout")
-    /// - `category.subcategory.name` (e.g., "network.peer.timeout")
-    ///
-    /// Requirements:
-    /// - Must contain at least one dot separator
-    /// - Each component must be non-empty
-    /// - Components should use lowercase_snake_case (not enforced, but recommended)
-    pub fn validate_id(id: &str) -> Result<(), ParameterValidationError> {
-        if id.is_empty() {
-            return Err(ParameterValidationError::InvalidParameterId {
-                id: id.to_string(),
-                reason: "parameter ID cannot be empty",
-            });
-        }
-
-        let parts: Vec<&str> = id.split('.').collect();
-        if parts.len() < 2 {
-            return Err(ParameterValidationError::InvalidParameterId {
-                id: id.to_string(),
-                reason: "parameter ID must follow 'category.name' format (missing dot separator)",
-            });
-        }
-
-        for (i, part) in parts.iter().enumerate() {
-            if part.is_empty() {
-                return Err(ParameterValidationError::InvalidParameterId {
-                    id: id.to_string(),
-                    reason: if i == 0 {
-                        "category component cannot be empty"
-                    } else if i == parts.len() - 1 {
-                        "name component cannot be empty"
-                    } else {
-                        "subcategory component cannot be empty"
-                    },
-                });
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Validate that a parameter ID uses a known category.
-    ///
-    /// This performs format validation via `validate_id()` and additionally
-    /// checks that the category is one of the known categories.
-    ///
-    /// Known categories: gossip, network, ledger, governance, trust, ratelimit, compute
-    ///
-    /// Use this for stricter validation when you want to ensure parameters
-    /// use only standard categories.
-    pub fn validate_known_category(id: &str) -> Result<(), ParameterValidationError> {
-        Self::validate_id(id)?;
-
-        let category = id.split('.').next().unwrap_or("");
-        if !KNOWN_PARAMETER_CATEGORIES.contains(&category) {
-            return Err(ParameterValidationError::InvalidParameterId {
-                id: id.to_string(),
-                reason: "unknown category (must be one of: gossip, network, ledger, governance, trust, ratelimit, compute)",
-            });
-        }
-
-        Ok(())
-    }
-
-    /// Create a new protocol parameter with ID validation.
-    ///
-    /// Returns an error if the ID doesn't follow the `category.name` format.
-    pub fn try_new(
-        id: impl Into<String>,
-        name: impl Into<String>,
-        description: impl Into<String>,
-        value: ParameterValue,
-    ) -> Result<Self, ParameterValidationError> {
-        let id_string = id.into();
-        Self::validate_id(&id_string)?;
-
-        Ok(Self {
-            id: id_string,
-            name: name.into(),
-            description: description.into(),
-            value,
-            constraints: ParameterConstraints::default(),
-            scope: ParameterScope::Global,
-            updated_at: icn_time::current_timestamp_secs(),
-            updated_by: None,
-            version: 0,
-        })
-    }
-
-    /// Create a new protocol parameter without ID validation.
-    ///
-    /// Prefer `try_new()` for user-provided IDs to ensure format compliance.
-    /// This method is useful for internal code where IDs are known to be valid.
-    pub fn new(
-        id: impl Into<String>,
-        name: impl Into<String>,
-        description: impl Into<String>,
-        value: ParameterValue,
-    ) -> Self {
-        Self {
-            id: id.into(),
-            name: name.into(),
-            description: description.into(),
-            value,
-            constraints: ParameterConstraints::default(),
-            scope: ParameterScope::Global,
-            updated_at: icn_time::current_timestamp_secs(),
-            updated_by: None,
-            version: 0,
-        }
-    }
-
-    /// Set the scope for this parameter
-    #[must_use]
-    pub fn with_scope(mut self, scope: ParameterScope) -> Self {
-        self.scope = scope;
-        self
-    }
-
-    /// Set constraints for this parameter
-    #[must_use]
-    pub fn with_constraints(mut self, constraints: ParameterConstraints) -> Self {
-        self.constraints = constraints;
-        self
-    }
-
-    /// Set minimum value constraint
-    #[must_use]
-    pub fn with_min(mut self, min: ParameterValue) -> Self {
-        self.constraints.min = Some(min);
-        self
-    }
-
-    /// Set maximum value constraint
-    #[must_use]
-    pub fn with_max(mut self, max: ParameterValue) -> Self {
-        self.constraints.max = Some(max);
-        self
-    }
-
-    /// Mark parameter as requiring restart to take effect
-    #[must_use]
-    pub fn requires_restart(mut self) -> Self {
-        self.constraints.requires_restart = true;
-        self
-    }
-
-    /// Validate a new value against constraints
-    pub fn validate(&self, new_value: &ParameterValue) -> Result<(), ParameterValidationError> {
-        // Check type compatibility
-        if std::mem::discriminant(&self.value) != std::mem::discriminant(new_value) {
-            return Err(ParameterValidationError::TypeMismatch {
-                expected: self.value.type_name(),
-                got: new_value.type_name(),
-            });
-        }
-
-        // Check for NaN/Infinity in floating-point values
-        match new_value {
-            ParameterValue::Float(f) if !f.is_finite() => {
-                return Err(ParameterValidationError::InvalidFloatValue {
-                    value: *f,
-                    reason: "value must be finite (not NaN or Infinity)",
-                });
-            }
-            ParameterValue::Percentage(f) if !f.is_finite() => {
-                return Err(ParameterValidationError::InvalidFloatValue {
-                    value: *f,
-                    reason: "percentage must be finite (not NaN or Infinity)",
-                });
-            }
-            // Check percentage range (0.0-100.0)
-            ParameterValue::Percentage(f) if !(*f >= 0.0 && *f <= 100.0) => {
-                return Err(ParameterValidationError::PercentageOutOfRange { value: *f });
-            }
-            _ => {}
-        }
-
-        // Check minimum constraint
-        if let Some(ref min) = self.constraints.min {
-            if new_value < min {
-                return Err(ParameterValidationError::BelowMinimum {
-                    value: new_value.clone(),
-                    min: min.clone(),
-                });
-            }
-        }
-
-        // Check maximum constraint
-        if let Some(ref max) = self.constraints.max {
-            if new_value > max {
-                return Err(ParameterValidationError::AboveMaximum {
-                    value: new_value.clone(),
-                    max: max.clone(),
-                });
-            }
-        }
-
-        // Check allowed values (using approximate equality for floats)
-        if let Some(ref allowed) = self.constraints.allowed_values {
-            let is_allowed = allowed.iter().any(|a| new_value.approximately_eq(a));
-            if !is_allowed {
-                return Err(ParameterValidationError::NotAllowed {
-                    value: new_value.clone(),
-                    allowed: allowed.clone(),
-                });
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Get the parameter category (first part of ID before the dot)
-    ///
-    /// Parameter IDs follow a dot-separated naming convention:
-    /// `<category>.<subcategory>.<name>` (e.g., "gossip.max_message_size")
-    ///
-    /// The category is the first component before the first dot.
-    /// If there is no dot, the entire ID is returned as the category.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// let param = ProtocolParameter::new("gossip.max_message_size", ...);
-    /// assert_eq!(param.category(), "gossip");
-    ///
-    /// let param = ProtocolParameter::new("simple_param", ...);
-    /// assert_eq!(param.category(), "simple_param");
-    /// ```
-    pub fn category(&self) -> &str {
-        self.id.split('.').next().unwrap_or(&self.id)
-    }
-}
-
-// ============================================================================
-// ParameterValue
-// ============================================================================
-
-/// Value types for protocol parameters
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "value")]
-pub enum ParameterValue {
-    /// Integer value (signed 64-bit)
-    Integer(i64),
-
-    /// Floating-point value
-    Float(f64),
-
-    /// String value
-    String(String),
-
-    /// Boolean value
-    Boolean(bool),
-
-    /// Duration in seconds
-    Duration(u64),
-
-    /// Size in bytes
-    Bytes(u64),
-
-    /// Percentage (0.0 to 100.0)
-    Percentage(f64),
-}
-
-impl ParameterValue {
-    /// Get the type name for error messages
-    pub fn type_name(&self) -> &'static str {
-        match self {
-            ParameterValue::Integer(_) => "integer",
-            ParameterValue::Float(_) => "float",
-            ParameterValue::String(_) => "string",
-            ParameterValue::Boolean(_) => "boolean",
-            ParameterValue::Duration(_) => "duration",
-            ParameterValue::Bytes(_) => "bytes",
-            ParameterValue::Percentage(_) => "percentage",
-        }
-    }
-
-    /// Get as integer, if applicable
-    pub fn as_integer(&self) -> Option<i64> {
-        match self {
-            ParameterValue::Integer(v) => Some(*v),
-            _ => None,
-        }
-    }
-
-    /// Get as float, if applicable
-    pub fn as_float(&self) -> Option<f64> {
-        match self {
-            ParameterValue::Float(v) => Some(*v),
-            ParameterValue::Percentage(v) => Some(*v),
-            _ => None,
-        }
-    }
-
-    /// Get as string, if applicable
-    pub fn as_string(&self) -> Option<&str> {
-        match self {
-            ParameterValue::String(v) => Some(v),
-            _ => None,
-        }
-    }
-
-    /// Get as boolean, if applicable
-    pub fn as_bool(&self) -> Option<bool> {
-        match self {
-            ParameterValue::Boolean(v) => Some(*v),
-            _ => None,
-        }
-    }
-
-    /// Get as duration in seconds, if applicable
-    pub fn as_duration_secs(&self) -> Option<u64> {
-        match self {
-            ParameterValue::Duration(v) => Some(*v),
-            _ => None,
-        }
-    }
-
-    /// Get as bytes, if applicable
-    pub fn as_bytes(&self) -> Option<u64> {
-        match self {
-            ParameterValue::Bytes(v) => Some(*v),
-            _ => None,
-        }
-    }
-}
-
-/// Epsilon for floating-point comparisons in governance parameters.
-///
-/// Uses 1e-6 (0.0001%) as a practical tolerance. The comparison uses
-/// relative epsilon for large values and absolute epsilon for small values:
-/// - epsilon = max(1e-6, value * 1e-6)
-///
-/// For a 66.67% threshold: epsilon = 66.67 * 1e-6 ≈ 0.00007 (7 millionths of a percent)
-/// For a 0.1% fee rate: epsilon = 1e-6 (absolute, not relative)
-///
-/// This handles JSON round-trip precision issues while maintaining sufficient
-/// precision for governance use cases.
-///
-/// IMPORTANT: For financial parameters requiring exact precision (e.g., fee rates
-/// in basis points, exchange rates), use Integer type instead of Float/Percentage.
-/// Example: Instead of Float(0.0025) for 0.25%, use Integer(25) for 25 basis points.
-const FLOAT_COMPARISON_EPSILON: f64 = 1e-6;
-
-impl ParameterValue {
-    /// Compare two floats with epsilon-based tolerance.
-    /// Returns true if the values are approximately equal.
-    fn floats_approximately_equal(a: f64, b: f64) -> bool {
-        if a == b {
-            return true; // Handles infinities and exact equality
-        }
-        let diff = (a - b).abs();
-        let max_abs = a.abs().max(b.abs());
-        // Use relative epsilon for large values, absolute epsilon for small values
-        diff <= FLOAT_COMPARISON_EPSILON.max(max_abs * FLOAT_COMPARISON_EPSILON)
-    }
-
-    /// Compare two floats with epsilon-based tolerance.
-    /// Returns Some(Ordering) based on approximate comparison.
-    fn float_partial_cmp(a: f64, b: f64) -> Option<std::cmp::Ordering> {
-        if Self::floats_approximately_equal(a, b) {
-            Some(std::cmp::Ordering::Equal)
-        } else {
-            a.partial_cmp(&b)
-        }
-    }
-
-    /// Check if this value is approximately equal to another (for float types).
-    /// For non-float types, uses exact equality.
-    pub fn approximately_eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (ParameterValue::Float(a), ParameterValue::Float(b)) => {
-                Self::floats_approximately_equal(*a, *b)
-            }
-            (ParameterValue::Percentage(a), ParameterValue::Percentage(b)) => {
-                Self::floats_approximately_equal(*a, *b)
-            }
-            // For non-float types, use exact equality
-            _ => self == other,
-        }
-    }
-}
-
-impl PartialOrd for ParameterValue {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        match (self, other) {
-            (ParameterValue::Integer(a), ParameterValue::Integer(b)) => a.partial_cmp(b),
-            (ParameterValue::Float(a), ParameterValue::Float(b)) => Self::float_partial_cmp(*a, *b),
-            (ParameterValue::Duration(a), ParameterValue::Duration(b)) => a.partial_cmp(b),
-            (ParameterValue::Bytes(a), ParameterValue::Bytes(b)) => a.partial_cmp(b),
-            (ParameterValue::Percentage(a), ParameterValue::Percentage(b)) => {
-                Self::float_partial_cmp(*a, *b)
-            }
-            (ParameterValue::String(a), ParameterValue::String(b)) => a.partial_cmp(b),
-            // Boolean and cross-type comparisons are not ordered
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for ParameterValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ParameterValue::Integer(v) => write!(f, "{v}"),
-            ParameterValue::Float(v) => write!(f, "{v}"),
-            ParameterValue::String(v) => write!(f, "\"{v}\""),
-            ParameterValue::Boolean(v) => write!(f, "{v}"),
-            ParameterValue::Duration(v) => {
-                if *v >= 86400 {
-                    write!(f, "{}d", v / 86400)
-                } else if *v >= 3600 {
-                    write!(f, "{}h", v / 3600)
-                } else if *v >= 60 {
-                    write!(f, "{}m", v / 60)
-                } else {
-                    write!(f, "{v}s")
-                }
-            }
-            ParameterValue::Bytes(v) => {
-                if *v >= 1_073_741_824 {
-                    write!(f, "{}GB", v / 1_073_741_824)
-                } else if *v >= 1_048_576 {
-                    write!(f, "{}MB", v / 1_048_576)
-                } else if *v >= 1024 {
-                    write!(f, "{}KB", v / 1024)
-                } else {
-                    write!(f, "{v}B")
-                }
-            }
-            ParameterValue::Percentage(v) => write!(f, "{v:.1}%"),
-        }
-    }
-}
-
-// ============================================================================
-// ParameterConstraints
-// ============================================================================
-
-/// Constraints on parameter values
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct ParameterConstraints {
-    /// Minimum allowed value (inclusive)
-    pub min: Option<ParameterValue>,
-
-    /// Maximum allowed value (inclusive)
-    pub max: Option<ParameterValue>,
-
-    /// If set, value must be one of these options
-    pub allowed_values: Option<Vec<ParameterValue>>,
-
-    /// Whether changing this parameter requires a daemon restart
-    pub requires_restart: bool,
-
-    /// Whether this parameter can be overridden at lower scopes
-    pub allow_override: bool,
-}
-
-impl ParameterConstraints {
-    /// Create constraints with min and max bounds
-    pub fn bounded(min: ParameterValue, max: ParameterValue) -> Self {
-        Self {
-            min: Some(min),
-            max: Some(max),
-            ..Default::default()
-        }
-    }
-
-    /// Create constraints with specific allowed values
-    pub fn enumerated(values: Vec<ParameterValue>) -> Self {
-        Self {
-            allowed_values: Some(values),
-            ..Default::default()
-        }
-    }
-}
-
-// ============================================================================
-// ParameterScope
-// ============================================================================
-
-/// Scope at which a parameter applies
-///
-/// Parameters cascade from Global -> Federation -> Cooperative,
-/// with more specific scopes overriding less specific ones.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type")]
-pub enum ParameterScope {
-    /// Applies to all nodes in the network
-    Global,
-
-    /// Applies to a specific federation
-    Federation {
-        /// Federation entity ID
-        id: EntityId,
-    },
-
-    /// Applies to a specific cooperative
-    Cooperative {
-        /// Cooperative entity ID
-        id: EntityId,
-    },
-}
-
-impl ParameterScope {
-    /// Check if this scope is more specific than another
-    pub fn is_more_specific_than(&self, other: &ParameterScope) -> bool {
-        match (self, other) {
-            // Cooperative is most specific
-            (ParameterScope::Cooperative { .. }, ParameterScope::Federation { .. }) => true,
-            (ParameterScope::Cooperative { .. }, ParameterScope::Global) => true,
-            // Federation is more specific than Global
-            (ParameterScope::Federation { .. }, ParameterScope::Global) => true,
-            // Same scope or less specific
-            _ => false,
-        }
-    }
-
-    /// Get the entity ID if this is an entity-scoped parameter
-    pub fn entity_id(&self) -> Option<&EntityId> {
-        match self {
-            ParameterScope::Global => None,
-            ParameterScope::Federation { id } => Some(id),
-            ParameterScope::Cooperative { id } => Some(id),
-        }
-    }
-}
-
-impl fmt::Display for ParameterScope {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ParameterScope::Global => write!(f, "global"),
-            ParameterScope::Federation { id } => write!(f, "federation:{}", id.identifier()),
-            ParameterScope::Cooperative { id } => write!(f, "cooperative:{}", id.identifier()),
-        }
-    }
-}
-
-// ============================================================================
-// ParameterChange (for history tracking)
-// ============================================================================
-
-/// Record of a parameter change for audit trail
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ParameterChange {
-    /// Parameter ID that was changed
-    pub parameter_id: String,
-
-    /// Previous value
-    pub old_value: ParameterValue,
-
-    /// New value
-    pub new_value: ParameterValue,
-
-    /// When the change occurred (Unix timestamp)
-    pub changed_at: u64,
-
-    /// Proposal ID that authorized this change (if any)
-    pub proposal_id: Option<String>,
-
-    /// DID of the actor who made the change
-    pub changed_by: Option<String>,
-}
+// Re-export all protocol parameter types from kernel-api.
+// These types are defined in kernel-api to allow icn-core to depend on them
+// without pulling in icn-governance.
+pub use icn_kernel_api::protocol_params::{
+    ParameterChange, ParameterConstraints, ParameterScope, ParameterValidationError,
+    ParameterValue, PendingChangeId, PendingChangeStatus, PendingParameterChange,
+    ProtocolParameter, ProtocolParameterStore, KNOWN_PARAMETER_CATEGORIES,
+};
 
 // ============================================================================
 // ProtocolChangeProposal
@@ -750,269 +127,6 @@ impl ProtocolChangeProposal {
     pub fn with_scope(mut self, scope: ParameterScope) -> Self {
         self.scope = Some(scope);
         self
-    }
-}
-
-// ============================================================================
-// ParameterValidationError
-// ============================================================================
-
-/// Errors that can occur when validating parameter values
-///
-/// These errors provide detailed information about validation failures,
-/// including the specific constraint that was violated and guidance for
-/// valid values.
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum ParameterValidationError {
-    /// Value type doesn't match parameter type
-    ///
-    /// Example: Setting an Integer parameter with a String value.
-    #[error("Type mismatch: expected {expected}, got {got}. Ensure the value type matches the parameter definition.")]
-    TypeMismatch {
-        /// Expected type
-        expected: &'static str,
-        /// Actual type
-        got: &'static str,
-    },
-
-    /// Value is below minimum
-    ///
-    /// Example: Setting `gossip.fanout` to 1 when minimum is 2.
-    #[error("Value {value} is below minimum {min}. Provide a value >= {min}.")]
-    BelowMinimum {
-        /// The value that was provided
-        value: ParameterValue,
-        /// The minimum allowed value
-        min: ParameterValue,
-    },
-
-    /// Value is above maximum
-    ///
-    /// Example: Setting `network.max_connections` to 10000 when maximum is 1000.
-    #[error("Value {value} is above maximum {max}. Provide a value <= {max}.")]
-    AboveMaximum {
-        /// The value that was provided
-        value: ParameterValue,
-        /// The maximum allowed value
-        max: ParameterValue,
-    },
-
-    /// Parameter ID format is invalid
-    ///
-    /// Valid format: `category.name` or `category.subcategory.name`
-    /// Examples: "gossip.fanout", "network.peer.timeout"
-    #[error(
-        "Invalid parameter ID '{id}': {reason}. Use format: category.name (e.g., 'gossip.fanout')."
-    )]
-    InvalidParameterId {
-        /// The invalid ID
-        id: String,
-        /// Why the ID is invalid
-        reason: &'static str,
-    },
-
-    /// Percentage value is out of valid range (0.0-100.0)
-    ///
-    /// Percentages are represented as 0.0 to 100.0, not as decimals.
-    #[error("Percentage {value} is out of range (must be 0.0-100.0). Use 50.0 for 50%, not 0.5.")]
-    PercentageOutOfRange {
-        /// The invalid percentage value
-        value: f64,
-    },
-
-    /// Value is not in allowed set
-    ///
-    /// Some parameters only accept specific values from a predefined list.
-    #[error("Value {value} is not in allowed set. Valid values: {allowed:?}")]
-    NotAllowed {
-        /// The value that was provided
-        value: ParameterValue,
-        /// The allowed values
-        allowed: Vec<ParameterValue>,
-    },
-
-    /// Invalid floating-point value (NaN or Infinity)
-    ///
-    /// Floating-point values must be finite numbers.
-    #[error("Invalid float value {value}: {reason}. Use a finite number.")]
-    InvalidFloatValue {
-        /// The invalid value
-        value: f64,
-        /// Reason for rejection
-        reason: &'static str,
-    },
-
-    /// Parameter not found
-    ///
-    /// The parameter ID does not exist in the parameter store.
-    #[error("Parameter not found: '{0}'. Use list_parameters() to see available parameters.")]
-    NotFound(String),
-
-    /// Parameter cannot be modified at this scope
-    ///
-    /// Some parameters can only be set at Global scope and cannot be overridden
-    /// by Federations or Cooperatives.
-    #[error("Parameter '{parameter_id}' cannot be overridden at {scope} scope. Check constraints.allow_override.")]
-    ScopeNotAllowed {
-        /// The parameter ID
-        parameter_id: String,
-        /// The scope that was attempted
-        scope: ParameterScope,
-    },
-}
-
-// ============================================================================
-// PendingParameterChange (for delayed execution)
-// ============================================================================
-
-/// Unique identifier for a pending parameter change
-pub type PendingChangeId = String;
-
-/// Status of a pending parameter change
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum PendingChangeStatus {
-    /// Change is scheduled but not yet applied
-    Pending,
-
-    /// Change was successfully applied at the effective time
-    Applied,
-
-    /// Change was cancelled before it took effect
-    Cancelled,
-
-    /// Change was superseded by a newer change to the same parameter
-    Superseded,
-}
-
-impl fmt::Display for PendingChangeStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            PendingChangeStatus::Pending => write!(f, "pending"),
-            PendingChangeStatus::Applied => write!(f, "applied"),
-            PendingChangeStatus::Cancelled => write!(f, "cancelled"),
-            PendingChangeStatus::Superseded => write!(f, "superseded"),
-        }
-    }
-}
-
-/// A scheduled parameter change waiting to take effect
-///
-/// When a protocol change proposal with `effective_at` is approved,
-/// a `PendingParameterChange` is created and stored. The background
-/// scheduler periodically checks for due changes and applies them.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PendingParameterChange {
-    /// Unique identifier for this pending change
-    pub id: PendingChangeId,
-
-    /// ID of the parameter being changed
-    pub parameter_id: String,
-
-    /// The new value to apply
-    pub new_value: ParameterValue,
-
-    /// Unix timestamp when this change should take effect
-    pub effective_at: u64,
-
-    /// Scope where this change applies
-    pub scope: ParameterScope,
-
-    /// ID of the proposal that authorized this change
-    pub proposal_id: String,
-
-    /// Unix timestamp when this pending change was created
-    pub created_at: u64,
-
-    /// Current status of this pending change
-    pub status: PendingChangeStatus,
-
-    /// Rationale for the change (copied from proposal)
-    pub rationale: String,
-
-    /// If superseded, the ID of the change that superseded this one
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub superseded_by: Option<PendingChangeId>,
-
-    /// If applied, the timestamp when it was applied
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub applied_at: Option<u64>,
-
-    /// If cancelled, the reason for cancellation
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cancellation_reason: Option<String>,
-}
-
-impl PendingParameterChange {
-    /// Create a new pending parameter change
-    pub fn new(
-        id: impl Into<String>,
-        parameter_id: impl Into<String>,
-        new_value: ParameterValue,
-        effective_at: u64,
-        scope: ParameterScope,
-        proposal_id: impl Into<String>,
-        rationale: impl Into<String>,
-    ) -> Self {
-        Self {
-            id: id.into(),
-            parameter_id: parameter_id.into(),
-            new_value,
-            effective_at,
-            scope,
-            proposal_id: proposal_id.into(),
-            created_at: icn_time::current_timestamp_secs(),
-            status: PendingChangeStatus::Pending,
-            rationale: rationale.into(),
-            superseded_by: None,
-            applied_at: None,
-            cancellation_reason: None,
-        }
-    }
-
-    /// Generate a unique ID for a pending change
-    ///
-    /// Format: `pending:{parameter_id}:{timestamp}:{counter}`
-    /// Uses an atomic counter for uniqueness when multiple changes are created in the same second.
-    pub fn generate_id(parameter_id: &str) -> String {
-        use std::sync::atomic::{AtomicU64, Ordering};
-        static COUNTER: AtomicU64 = AtomicU64::new(0);
-
-        let timestamp = icn_time::current_timestamp_secs();
-        let counter = COUNTER.fetch_add(1, Ordering::Relaxed);
-        format!("pending:{parameter_id}:{timestamp}:{counter:08x}")
-    }
-
-    /// Check if this change is due (effective_at <= current time)
-    pub fn is_due(&self, current_time: u64) -> bool {
-        self.status == PendingChangeStatus::Pending && self.effective_at <= current_time
-    }
-
-    /// Mark this change as applied
-    pub fn mark_applied(&mut self) {
-        self.status = PendingChangeStatus::Applied;
-        self.applied_at = Some(icn_time::current_timestamp_secs());
-    }
-
-    /// Mark this change as cancelled
-    pub fn mark_cancelled(&mut self, reason: impl Into<String>) {
-        self.status = PendingChangeStatus::Cancelled;
-        self.cancellation_reason = Some(reason.into());
-    }
-
-    /// Mark this change as superseded by another change
-    pub fn mark_superseded(&mut self, superseded_by: impl Into<String>) {
-        self.status = PendingChangeStatus::Superseded;
-        self.superseded_by = Some(superseded_by.into());
-    }
-
-    /// Time until this change takes effect (in seconds), or None if already due
-    pub fn time_until_effective(&self, current_time: u64) -> Option<u64> {
-        if self.effective_at > current_time {
-            Some(self.effective_at - current_time)
-        } else {
-            None
-        }
     }
 }
 
@@ -1364,10 +478,10 @@ mod tests {
     fn test_scope_specificity() {
         let global = ParameterScope::Global;
         let fed = ParameterScope::Federation {
-            id: EntityId::federation("test-fed").unwrap(),
+            id: "test-fed".to_string(),
         };
         let coop = ParameterScope::Cooperative {
-            id: EntityId::cooperative("test-coop").unwrap(),
+            id: "test-coop".to_string(),
         };
 
         assert!(fed.is_more_specific_than(&global));
@@ -1382,12 +496,12 @@ mod tests {
         assert_eq!(ParameterScope::Global.to_string(), "global");
 
         let fed = ParameterScope::Federation {
-            id: EntityId::federation("midwest-fed").unwrap(),
+            id: "midwest-fed".to_string(),
         };
         assert_eq!(fed.to_string(), "federation:midwest-fed");
 
         let coop = ParameterScope::Cooperative {
-            id: EntityId::cooperative("food-coop").unwrap(),
+            id: "food-coop".to_string(),
         };
         assert_eq!(coop.to_string(), "cooperative:food-coop");
     }
