+++ icn/crates/icn-kernel-api/Cargo.toml
@@ -10,6 +10,7 @@ keywords = ["icn", "kernel", "api", "traits"]
 categories = ["api-bindings"]
 
 [dependencies]
+anyhow = "1.0"
 thiserror = "2.0"
 serde = { version = "1.0", features = ["derive"] }
 arc-swap = "1.8"
